requirements.txt
adal==1.2.7
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
applicationinsights==0.11.10
argcomplete==3.5.3
azure-appconfiguration==1.7.1
azure-batch==15.0.0b2
azure-cli==2.73.0
azure-cli-core==2.73.0
azure-cli-telemetry==1.1.0
azure-common==1.1.28
azure-core==1.34.0
azure-cosmos==3.2.0
azure-data-tables==12.4.0
azure-datalake-store==1.0.0a0
azure-functions==1.23.0
azure-keyvault-administration==4.4.0b2
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.11.0b1
azure-keyvault-secrets==4.7.0
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==4.0.0
azure-mgmt-appconfiguration==3.1.0
azure-mgmt-appcontainers==2.0.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==4.0.0
azure-mgmt-batch==17.3.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==2.0.0
azure-mgmt-cdn==12.0.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-compute==34.1.0
azure-mgmt-containerinstance==10.2.0b1
azure-mgmt-containerregistry==14.0.0
azure-mgmt-containerservice==35.0.0
azure-mgmt-core==1.5.0
azure-mgmt-cosmosdb==9.7.0
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datalake-store==1.1.0b1
azure-mgmt-datamigration==10.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==10.2.0b2
azure-mgmt-eventhub==10.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hdinsight==9.0.0b3
azure-mgmt-imagebuilder==1.3.0
azure-mgmt-iotcentral==10.0.0b2
azure-mgmt-iothub==3.0.0
azure-mgmt-iothubprovisioningservices==1.1.0
azure-mgmt-keyvault==11.0.0
azure-mgmt-loganalytics==13.0.0b4
azure-mgmt-managementgroups==1.0.0
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==9.0.0
azure-mgmt-monitor==7.0.0b1
azure-mgmt-msi==7.0.0
azure-mgmt-mysqlflexibleservers==1.0.0b3
azure-mgmt-netapp==10.1.0
azure-mgmt-policyinsights==1.1.0b4
azure-mgmt-postgresqlflexibleservers==1.1.0b2
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==10.2.0b17
azure-mgmt-recoveryservices==3.0.0
azure-mgmt-recoveryservicesbackup==9.1.0
azure-mgmt-redhatopenshift==1.5.0
azure-mgmt-redis==14.5.0
azure-mgmt-resource==23.3.0
azure-mgmt-search==9.1.0
azure-mgmt-security==6.0.0
azure-mgmt-servicebus==8.2.1
azure-mgmt-servicefabric==2.1.0
azure-mgmt-servicefabricmanagedclusters==2.1.0b1
azure-mgmt-servicelinker==1.2.0b3
azure-mgmt-signalr==2.0.0b2
azure-mgmt-sql==4.0.0b21
azure-mgmt-sqlvirtualmachine==1.0.0b5
azure-mgmt-storage==22.1.0
azure-mgmt-synapse==2.1.0b5
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==7.3.1
azure-monitor-query==1.2.0
azure-multiapi-storage==1.4.0
azure-storage-blob==12.25.1
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.20.0
azure-synapse-managedprivateendpoints==0.4.0
azure-synapse-spark==0.7.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.2
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.0
fabric==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
invoke==2.2.0
isodate==0.7.2
javaproperties==0.5.2
jiter==0.10.0
jmespath==1.0.1
jsondiff==2.0.0
knack==0.11.0
MarkupSafe==3.0.2
microsoft-security-utilities-secret-masker==1.0.0b4
msal==1.32.3
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
numpy==2.2.6
oauthlib==3.2.2
openai==1.79.0
packaging==25.0
paramiko==3.5.1
pkginfo==1.12.1.2
portalocker==2.10.1
psutil==7.0.0
py-deviceid==0.1.1
pycomposefile==0.0.32
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
PyGithub==1.59.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
scp==0.13.6
semver==2.13.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
sshtunnel==0.1.5
tabulate==0.9.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.3.3
Werkzeug==3.1.3
wrapt==1.17.2
xmltodict==0.14.2
